Refactor Brief and CartContext components for improved performance and code organization
